+++ doc/changes.rst
@@ -7,6 +7,9 @@ New in 0.33
 -----------
 
 * API :func:`load.fiff.events`:  The merge parameter is now by default inferred based on the raw data
+* Experiment :mod:`pipeline`:
+
+  - New :atr:`MneExperiment.merge_events` attribute
 
 
 New in 0.32